requirements.txt
alembic==1.8.1
anyio==3.6.1
click==8.1.3
colorama==0.4.5
fastapi==0.83.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
Mako==1.2.2
MarkupSafe==2.1.1
psycopg2==2.9.3
pydantic==1.10.2
python-decouple==3.6
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
